fix: home페이지 요소와 location페이지 요소를 뒤바꾸고 그에따른 수정
