+++ src/App.tsx
@@ -9,6 +9,9 @@ import PwdRecovery from "./pages/PwdRecovery";
 const Home = lazy(() => import("./pages/Home"));
 const Login = lazy(() => import("./pages/Login"));
 const Register = lazy(() => import("./pages/Register"));
+const Write = lazy(() => import("./pages/Write"));
+const Mypage = lazy(() => import("./pages/Mypage"));
+const Location = lazy(() => import("./pages/Location"));
 
 function App() {
   return (
@@ -17,7 +20,9 @@ function App() {
         <Routes>
           <Route path="/" element={<BottomNavigationLayout />}>
             <Route index element={<Home />} />
-            <Route path="/about" element={<Home />} />
+            <Route path="/location" element={<Location />} />
+            <Route path="/write" element={<Write />} />
+            <Route path="/mypage" element={<Mypage />} />
           </Route>
           <Route path="/login" element={<BottomButtonLayout />}>
             <Route index element={<Login />} />
